dev to main (#58)
* Iframe (#55)

* Iframe

Try if we can use an iframe to fetch the roadmap from confluence

* Simple roadmap.md

Just a table is good enough for now

* remove roadmap from website

We don't want the roadmap on the website.
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* add roadmap to website

We want the roadmap on the website. So a link will be enough
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* Roadmap on docswebsite (#57)

* Iframe

Try if we can use an iframe to fetch the roadmap from confluence

* Simple roadmap.md

Just a table is good enough for now

* remove roadmap from website

We don't want the roadmap on the website.
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* add roadmap to website

We want the roadmap on the website. So a link will be enough
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* Nice text for the roadmap

* add picture and change order of UU

* remove samples, and move the code to guides

* remove samples, and move the code to guides

* remove samples, and move the code to guides, indendation

* welcome to the guides

* clean up

* just start :-)

* have all the links open in a new browser

* picture of openshift

* Roadmap on docswebsite (#59)

* Iframe

Try if we can use an iframe to fetch the roadmap from confluence

* Simple roadmap.md

Just a table is good enough for now

* remove roadmap from website

We don't want the roadmap on the website.
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* add roadmap to website

We want the roadmap on the website. So a link will be enough
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* Nice text for the roadmap

* add picture and change order of UU

* remove samples, and move the code to guides

* remove samples, and move the code to guides

* remove samples, and move the code to guides, indendation

* welcome to the guides

* clean up

* just start :-)

* have all the links open in a new browser

* picture of openshift

* add picture landing