Fix tls guide (#64)
* dev to main (#63)

* Iframe (#55)

* Iframe

Try if we can use an iframe to fetch the roadmap from confluence

* Simple roadmap.md

Just a table is good enough for now

* remove roadmap from website

We don't want the roadmap on the website.
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* add roadmap to website

We want the roadmap on the website. So a link will be enough
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* Roadmap on docswebsite (#57)

* Iframe

Try if we can use an iframe to fetch the roadmap from confluence

* Simple roadmap.md

Just a table is good enough for now

* remove roadmap from website

We don't want the roadmap on the website.
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* add roadmap to website

We want the roadmap on the website. So a link will be enough
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* Nice text for the roadmap

* add picture and change order of UU

* remove samples, and move the code to guides

* remove samples, and move the code to guides

* remove samples, and move the code to guides, indendation

* welcome to the guides

* clean up

* just start :-)

* have all the links open in a new browser

* picture of openshift

* Roadmap on docswebsite (#59)

* Iframe

Try if we can use an iframe to fetch the roadmap from confluence

* Simple roadmap.md

Just a table is good enough for now

* remove roadmap from website

We don't want the roadmap on the website.
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* add roadmap to website

We want the roadmap on the website. So a link will be enough
We will have 1 source of truth for the roadmap: https://utrechtuniversity.atlassian.net/wiki/spaces/LNX/pages/edit-v2/613842961

* Nice text for the roadmap

* add picture and change order of UU

* remove samples, and move the code to guides

* remove samples, and move the code to guides

* remove samples, and move the code to guides, indendation

* welcome to the guides

* clean up

* just start :-)

* have all the links open in a new browser

* picture of openshift

* add picture landing

* feat: add postgresql on netapp documentation

* update libcrypto error

* correct keypath

* use key ed_25519

* use key ed_25519 also in path

* update prd and acc to use ed_25519

* dir .ssh instead of /tmp/.ssh

* kiss

* 400

* try .pem

* Added deploy-artifactory-chart guide (#62)

* Added deploy-artifactory-chart guide

* Enable https and force http to https

* Update deploy-artifactory-chart.md

run

* Update deploy-artifactory-chart.md

* Update deploy-artifactory-chart.md

to kickoff pipeline

* Update deploy-artifactory-chart.md

* Update deploy-artifactory-chart.md

* Update deploy-artifactory-chart.md

* Update deploy-artifactory-chart.md

* Update deploy-artifactory-chart.md

* Update deploy-artifactory-chart.md

---------

Co-authored-by: Joris Jansen <joris@tioli.xs4all.nl>
Co-authored-by: Rene <50985012+Rene-Universiteit-Utrecht@users.noreply.github.com>

* use rsa again.
Also the pipeline doesn't run from a fork

* update docs and tell the linux team how to work with this repo

---------

Co-authored-by: Christian Boon <c.p.j.boon@uu.nl>
Co-authored-by: Christian Boon <cpjboon@gmail.com>
Co-authored-by: Joris Jansen <102958224+JorisJansenUU@users.noreply.github.com>
Co-authored-by: Joris Jansen <joris@tioli.xs4all.nl>

* Fix file name

---------

Co-authored-by: Rene <50985012+Rene-Universiteit-Utrecht@users.noreply.github.com>
Co-authored-by: Christian Boon <c.p.j.boon@uu.nl>
Co-authored-by: Christian Boon <cpjboon@gmail.com>
Co-authored-by: Joris Jansen <102958224+JorisJansenUU@users.noreply.github.com>
Co-authored-by: Joris Jansen <joris@tioli.xs4all.nl>